+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+12-3-22 mark@coder.com
+1. added a Google Cloud Windows and Visual Studio template which prompts user for WinOS and VS release
+
 11-25-22 mark@coder.com
 1. added another docker-in-pod template using the updated Kubernetes Terraform provider.